산지조직 정산관리 수정

+++ src/main/java/com/at/apcss/pd/sprt/vo/SprtBizClclnDmndDtlVO.java
@@ -41,7 +41,7 @@ public class SprtBizClclnDmndDtlVO extends ComVO {
     /**
      * 요청금액
      */
-    private int dmndAmt;
+    private long dmndAmt;
 
     /**
      * 요청항목종류
